feat: rename the component to "Product Relations" and extract it from "src" folder due to new convention
