+++ apps/storefront-e2e/src/support/page_fragments/product-references.fragment.ts
@@ -1,4 +1,4 @@
 export class ProductReferencesFragment {
-  getWrapper = () => cy.get('oryx-product-references');
+  getWrapper = () => cy.get('oryx-product-relations');
   getProducts = () => this.getWrapper().find('oryx-product-card');
 }
