--- libs/domain/product/references/src/index.ts
@@ -1,2 +1,0 @@
-export * from './references.component';
-export * from './references.schema';
