+++ libs/domain/product/relations/index.ts
@@ -1,0 +1,2 @@
+export * from './relations.component';
+export * from './relations.schema';
