+++ libs/domain/product/relations/relations.component.ts
